Introduce URLParser
